feature/add-neoletter-tracking-id: move trackig logic to seperate file
